+++ crates/emmylua_code_analysis/resources/std/builtin.lua
@@ -181,11 +181,10 @@
 --- Language Server Optimization Items.
 ---
 --- Parameters:
---- - `skip_table_fields_check`: Skip table field diagnostics. It is recommended to use this option for all large configuration tables.
 --- - `delayed_definition`: Indicates that the type of the variable is determined by the first assignment.
 ---    Only valid for `local` declarations with no initial value.
 --- @class lsp_optimization: Attribute
---- @overload fun(code: "skip_table_fields_check"|"delayed_definition")
+--- @overload fun(code: "delayed_definition")
 
 ---
 --- Index field alias, will be displayed in `hint` and `completion`.